% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnoaa-package.R
\docType{package}
\name{readnoaa-package}
\alias{readnoaa}
\alias{readnoaa-package}
\title{readnoaa: Access 'NOAA' Climate and Weather Data}
\description{
Provides clean, tidy access to climate and weather data from the 'National Oceanic and Atmospheric Administration' ('NOAA') via the 'National Centers for Environmental Information' ('NCEI') Data Service API \url{https://www.ncei.noaa.gov/support/access-data-service-api-user-documentation}. Covers daily weather observations, monthly and annual summaries, and 30-year climate normals from over 100,000 stations across 180 countries. No API key is required. Dedicated functions handle the most common datasets, while a generic fetcher provides access to all 'NCEI' datasets. Station discovery functions help users find stations by location or name. Data is downloaded on first use and cached locally for subsequent calls. This package is not endorsed or certified by 'NOAA'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://charlescoverdale.github.io/readnoaa/}
  \item \url{https://github.com/charlescoverdale/readnoaa}
  \item Report bugs at \url{https://github.com/charlescoverdale/readnoaa/issues}
}

}
\author{
\strong{Maintainer}: Charles Coverdale \email{charlesfcoverdale@gmail.com} [copyright holder]

}
\concept{NCEI}
\concept{NOAA}
\concept{climate}
\concept{climate normals}
\concept{precipitation}
\concept{temperature}
\concept{weather}
\concept{weather stations}
\keyword{internal}
